Remove circular dependency in fluent-langneg
The circular dependency caused fluent-langneg to not be usable in Vite
and likely other bundlers. It also caused a warning to show up in Rollup.

fixes #459
